requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
async-timeout==5.0.1
asyncer==0.0.10
attrs==25.4.0
bitsandbytes==0.45.5
brotli==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
controlnet_aux==0.0.10
cycler==0.12.1
datasets==4.3.0
debugpy==1.8.18
decorator==5.2.1
diffusers==0.35.2
dill==0.4.0
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.122.0
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gradio==6.0.1
gradio_client==2.0.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMatting==1.1.14
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
rembg==2.0.68
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
terminado==0.18.1
tifffile==2025.5.10
timm==1.0.22
tokenizers==0.22.1
tomlkit==0.13.3
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.22.3
watchdog==6.0.0
wcwidth==0.2.14
xformers==0.0.29.post3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
